Updated mkdir and setup script. Checks run faster now.

+++ packages/phpunit.sh
@@ -1,3 +1,6 @@
 #!/bin/bash
 
-wget -q https://phar.phpunit.de/phpunit.phar -O /usr/local/bin/phpunit && chmod +x /usr/local/bin/phpunit
+if [[ ! -e /usr/local/bin/phpunit ]]; then
+	wget -q https://phar.phpunit.de/phpunit.phar -O /usr/local/bin/phpunit && chmod +x /usr/local/bin/phpunit
+	echo "/usr/local/bin/phpunit"
+fi
